// client/bookStorage.Spa/src/api/generated/dtos/categoryDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * BookStorage API
 * API for managing book storage
 * OpenAPI spec version: v1
 */

export interface CategoryDto {
  /** @nullable */
  id?: string | null;
  name: string;
  /** @nullable */
  parentCategoryId?: string | null;
  /** @nullable */
  subCategories?: CategoryDto[] | null;
}